buddy_mvp/requirements.txt
winrt-runtime==3.2.1
winrt-Windows.Foundation==3.2.1
winrt-Windows.Media.Ocr==3.2.1
pywin32==309 ; platform_system=="Windows"
Pillow==11.2.1
win10toast==0.9 ; platform_system=="Windows"
winrt-Windows.Graphics.Imaging==3.2.1
winrt-Windows.Storage.Streams==3.2.1
imagehash==4.3.1
numpy>=1.24.0
openai>=1.3.8
tiktoken>=0.6.0
python-dotenv>=1.0.1
